+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 1.0.1 / 2015-01-11
+------------------
 
 1. Fixes for `gulp-stylus` that broke backward compatibility, sourcemaps
    now provides by `gulp-sourcemaps` (as plugin).